Puts functions in private methods
Puts port handling in private method of Signal. Thus, in the main draw() loop there is only one call to Signal#get_data(). This is as it should be, because the technical details should be handled by Signal, not by the main function. (Even though version 1.5 of Processing seems to ignore the private key word...)